Rename decode.py to stream.py
