+++ sherpa/bin/conv_emformer_transducer_stateless2/beam_search.py
@@ -2,7 +2,7 @@ from typing import List
 
 import k2
 import torch
-from decode import Stream, stack_states, unstack_states
+from stream import Stream, stack_states, unstack_states
 
 from sherpa import fast_beam_search_one_best, streaming_greedy_search
