+++ sherpa/bin/conv_emformer_transducer_stateless2/streaming_server.py
@@ -41,7 +41,7 @@ import sentencepiece as spm
 import torch
 import websockets
 from beam_search import FastBeamSearch, GreedySearch
-from decode import Stream
+from stream import Stream
 
 from sherpa import RnntConvEmformerModel
