+++ sherpa/bin/streaming_pruned_transducer_statelessX/streaming_server.py
@@ -43,7 +43,7 @@ import sentencepiece as spm
 import torch
 import websockets
 from beam_search import FastBeamSearch, GreedySearch
-from decode import Stream
+from stream import Stream
 
 from sherpa import RnntConformerModel
